% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{checkbox_names}
\alias{checkbox_names}
\title{Change Checkboxes Names to Option Names}
\usage{
checkbox_names(data, dic, labels, checkbox_labels = c("No", "Yes"))
}
\arguments{
\item{data}{Dataset containing the REDCap data.}

\item{dic}{Dataset containing the REDCap dictionary.}

\item{labels}{Named character vector with the names of the variables in the data and their corresponding REDCap labels.}

\item{checkbox_labels}{Character vector specifying the names for the two options of each checkbox variable. The default is `c('No', 'Yes')`.}
}
\description{
This function updates the names of checkboxes in the dataset and dictionary to reflect the names of their options.
}
