% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{dades_events}
\alias{dades_events}
\title{Creation of a data frame with variables of all the forms of a specified event}
\usage{
dades_events(data, dic, event, which = NULL)
}
\arguments{
\item{data}{Preprocessed data.}

\item{dic}{Preprocessed dictionary.}

\item{event}{Downloaded instrument-event mapping from REDCap.}

\item{which}{Specify an event if only data for the desired event is wanted.}
}
\description{
Function that given the data, the dictionary and the mapping between forms and events it creates a dataset with the variables of all the forms that are in this event. It can be chosen to return only the data from the specified event.
}
