% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{dic_checkboxes}
\alias{dic_checkboxes}
\title{Change the names of checkboxes variables in the REDCap dictionary}
\usage{
dic_checkboxes(var_check, dic, labels, checkbox_labels = c("No", "Yes"))
}
\arguments{
\item{var_check}{Character vector containing the names of those variables that are checkboxes.}

\item{dic}{Dataset containing the REDCap dictionary.}

\item{labels}{Named character vector with the name of the variables in the data and the REDCap label in its name.}

\item{checkbox_labels}{Character vector with the names that will have the two options of every checkbox variable. Default is c('No', 'Yes').}
}
\description{
Auxiliary function to checkbox_names. Adds to the dictionary all variables that correspond to all the options of checkbox (with the name as it is in the data) and remove the original general checkbox variable.
}
