% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_dictionary.R
\name{rd_dictionary}
\alias{rd_dictionary}
\title{Transform the data dictionary and handle branching logic}
\usage{
rd_dictionary(project = NULL, data = NULL, dic = NULL, event_form = NULL)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{data}{A \code{data.frame} or \code{tibble} representing the REDCap dataset containing the checkbox variables.}

\item{dic}{A \code{data.frame} representing the REDCap dictionary with metadata, including field names, field types, and branching logic.}

\item{event_form}{A \code{data.frame} or \code{list} mapping event names to forms for longitudinal projects. Optional; defaults to \code{NULL} if not applicable.}
}
\value{
A list containing the following elements:
\item{data}{The original dataset, passed to the function.}
\item{dictionary}{The updated data dictionary, with modified branching logic.}
\item{event_form}{The original event-form mapping, passed to the function (if applicable).}
\item{results}{A string summarizing the results of the transformation process, including any variables with unconverted branching logic.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function updates the data dictionary by evaluating and transforming the branching logic expressions for each field in the dictionary.
It checks if any branching logic is present and attempts to convert it using the specified data, dictionary, and event-form mapping.
If there are any issues with the conversion, those fields are listed in the results.
}
\examples{

result <- covican |> rd_dictionary()

print(result$results)

}
