% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_export.R
\name{rd_export}
\alias{rd_export}
\title{Export Queries to an Excel File}
\usage{
rd_export(
  project = NULL,
  queries = NULL,
  column = NULL,
  sheet_name = NULL,
  path = NULL,
  password = NULL
)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{queries}{A data frame containing the identified queries. If \code{...} is provided, this argument is ignored.}

\item{column}{A string specifying the column in the dataset that contains hyperlinks. If not specified,
hyperlinks will not be added unless a column named \code{Link} is detected.}

\item{sheet_name}{A string specifying the name of the sheet in the resulting \code{.xlsx} file. Defaults to \code{"Sheet1"}.}

\item{path}{A string specifying the file path to save the \code{.xlsx} file. If \code{NULL}, the file is saved in the
current working directory with the name \code{example.xlsx}.}

\item{password}{An optional string to password-protect the worksheet, preventing unauthorized edits.}
}
\value{
An \code{.xlsx} file saved to the specified path, containing the query data and hyperlinks if specified.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function exports a query dataset, typically generated using \code{rd_query} or \code{rd_event}, into an \code{.xlsx} file.
It supports adding hyperlinks to specified columns and optional password protection for the worksheet.
}
\examples{
\dontrun{
# Export queries to an Excel file
rd_export(
  queries = my_queries,
  column = "Link",
  sheet_name = "My Queries",
  path = "queries.xlsx"
)
}

}
