% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_factor.R
\name{rd_factor}
\alias{rd_factor}
\title{Convert Variables to Factors in a REDCap Dataset}
\usage{
rd_factor(
  project = NULL,
  data = NULL,
  dic = NULL,
  event_form = NULL,
  exclude = NULL
)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{data}{A \code{data.frame} or \code{tibble} representing the REDCap dataset containing the checkbox variables.}

\item{dic}{A \code{data.frame} representing the REDCap dictionary with metadata, including field names, field types, and branching logic.}

\item{event_form}{A \code{data.frame} or \code{list} mapping event names to forms for longitudinal projects. Optional; defaults to \code{NULL} if not applicable.}

\item{exclude}{A character vector of variable names to exclude from being converted into factors.}
}
\value{
A list containing:
\item{data}{The transformed dataset with factor variables applied.}
\item{dictionary}{The dictionary with updated branching logic for the transformed variables.}
\item{event_form}{The event-form mapping used (if provided).}
\item{results}{A string summarizing the changes made during the transformation.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function converts variables in a REDCap dataset that have associated \code{.factor} columns into actual factor variables, while also updating branching logic in the associated dictionary.

It also allows for the exclusion of specific variables from being converted into factors.
}
\details{
This function searches for columns in the data that have a \code{.factor} suffix (indicating that they can be converted into factors) and converts them into factors based on their labels.
The \code{exclude} argument allows you to specify which variables should not be converted.
The function also modifies the branching logic in the dictionary to reflect the changes made in the data.

Variables with the names \code{redcap_event_name.factor} and \code{redcap_data_access_group.factor} are excluded from the conversion process to avoid altering event and access group information.
}
\examples{
result <- REDCapDM::rd_factor(covican, exclude = c("available_analytics", "urine_culture"))

transformed_data <- result$data

}
