% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{recalculate}
\alias{recalculate}
\title{Recalculate REDCap calculated fields}
\usage{
recalculate(data, dic)
}
\arguments{
\item{data}{Dataset containing the REDCap data.}

\item{dic}{Dataset containing the REDCap dictionary.}
}
\description{
Function that recalculates every calculated field if the logic can be transcribed to R. Recall that calculated fields with smart-variables in the logic or variables in other events cannot be transcribed.

The function will return the dataset and dictionary with the added recalculated variables (the name of the calculated field + `_recalc`) along with a table that shows the summary of the results.
}
