% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-suplement.R
\name{round}
\alias{round}
\title{Round Numbers to a Specified Number of Digits ----}
\usage{
round(x, digits)
}
\arguments{
\item{x}{A numeric vector to be rounded.}

\item{digits}{Integer indicating the number of decimal places to round to.}
}
\value{
A numeric vector rounded to the specified number of digits.
}
\description{
This function rounds numeric values to the specified number of decimal digits,
mimicking the behavior of the base R \code{round()} function but implemented manually.
}
\examples{
round(3.14159, 2)
round(c(-2.718, 3.14159), 1)

}
