% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{split_event}
\alias{split_event}
\title{Creation of a Data Frame with Variables from All Forms of a Specified Event}
\usage{
split_event(data, dic, event_form, which = NULL)
}
\arguments{
\item{data}{Data frame containing data from REDCap.}

\item{dic}{Data frame  containing the dictionary read from REDCap.}

\item{event_form}{Data frame containing the correspondence of each event with each form.}

\item{which}{Character string specifying an event if only data for that event is desired.}
}
\description{
This function generates a nested dataset filtered by each event, containing only the variables associated with each event. It uses the provided data, dictionary, and event-form mapping. You can choose to return data for a specific event.
}
