% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{split_form}
\alias{split_form}
\title{Creation of a data frame with variables of a specified form}
\usage{
split_form(data, dic, event, which = NULL, wide = FALSE)
}
\arguments{
\item{data}{Preprocessed data.}

\item{dic}{Preprocessed dictionary.}

\item{event}{Downloaded instrument-event mapping from REDCap.}

\item{which}{Specify a form if only data for the desired form is wanted.}

\item{wide}{If the dataset needs to be in a wide format or not (long).}
}
\description{
Function that given the data, the dictionary and the mapping between forms and events it creates a dataset with the variables that are in this form for all events. It can be chosen to return only the data from the specified form.
}
