% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{to_factor}
\alias{to_factor}
\title{Convert variables to factors}
\usage{
to_factor(data, exclude = NULL)
}
\arguments{
\item{data}{Dataset containing the REDCap data.}

\item{exclude}{Character vector containing the names of those variables that will not be converted to factors. If `NULL`, all variables will be converted.}
}
\description{
Function that converts every variable except those specified to factor.
}
