% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{transform_name}
\alias{transform_name}
\title{Auxiliary function to `checkbox_names`}
\usage{
transform_name(var_check, name, labels)
}
\arguments{
\item{var_check}{a character vector containing the names of those variables that are checkboxes}

\item{name}{a character element with the original name of the checkbox variable that has to be changed}

\item{labels}{a named character vector with the name of the variables in the data and the REDCap label in its name}
}
\description{
Auxiliary function to checkbox_names. It changes the name of the checkbox variable to the name of the option it corresponds
}
