% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_queries.R
\name{check_queries}
\alias{check_queries}
\title{Check modifications between two dataset of queries}
\usage{
check_queries(old, new)
}
\arguments{
\item{old}{Old version of the dataset of queries.}

\item{new}{New version of the dataset of queries. This object will be used to determine the status of each query.}
}
\value{
A data frame with a merge of all queries plus a column indicating which queries were resolved or are new comparing to the old queries dataset.
}
\description{
This function compares a former dataset of queries with a new one and allows you to check which of the old queries were resolved, which are yet to be resolved and which of them are new.
}
\keyword{check,}
\keyword{queries}
