% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{checkbox_names}
\alias{checkbox_names}
\title{Change checkboxes names into the name of their options}
\usage{
checkbox_names(data, dic, labels, checkbox_labels = c("No", "Yes"))
}
\arguments{
\item{data}{Dataset containing the REDCap data.}

\item{dic}{Dataset containing the REDCap dictionary.}

\item{labels}{Named character vector with the name of the variables in the data and the REDCap label in its name.}

\item{checkbox_labels}{Character vector with the names that will have the two options of every checkbox variable. Default is c('No', 'Yes').}
}
\description{
Function that returns both data and dictionary with the name of the checkboxes transformed by the name of their options.
}
