% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_checkbox.R
\name{rd_checkbox}
\alias{rd_checkbox}
\title{Transform Checkbox Variables in a REDCap Project}
\usage{
rd_checkbox(
  project = NULL,
  data = NULL,
  dic = NULL,
  event_form = NULL,
  checkbox_labels = c("No", "Yes"),
  checkbox_na = FALSE,
  checkbox_names = TRUE
)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{data}{A \code{data.frame} or \code{tibble} representing the REDCap dataset containing the checkbox variables.}

\item{dic}{A \code{data.frame} representing the REDCap dictionary with metadata, including field names, field types, and branching logic.}

\item{event_form}{A \code{data.frame} or \code{list} mapping event names to forms for longitudinal projects. Optional; defaults to \code{NULL} if not applicable.}

\item{checkbox_labels}{A character vector of length 2 specifying the labels to be used for the checkbox options. Defaults to \code{c("No", "Yes")}.}

\item{checkbox_na}{Logical indicating whether to assign \code{NA} to checkbox fields when the branching logic condition is not satisfied. Defaults to \code{FALSE}.}

\item{checkbox_names}{Logical indicating whether to rename the checkbox variables in the dataset and dictionary according to their label options. Defaults to \code{TRUE}.}
}
\value{
A list containing the following elements:
\item{data}{The transformed dataset with checkbox variables updated.}
\item{dictionary}{The updated dictionary reflecting any changes made to the checkbox fields, including renamed variables.}
\item{event_form}{The event-form mapping (if provided).}
\item{results}{A summary of the transformation process, including any issues with branching logic or fields that need review.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is used to convert checkbox variables in a REDCap dataset from their default categories (e.g., "Checked" and "Unchecked") to numeric values (0 and 1), and optionally, relabel and rename them according to user-defined options. It also evaluates branching logic for checkbox fields and adjusts the data and dictionary accordingly.
}
\details{
This function is primarily used to process checkbox fields in a REDCap project. It performs the following:
\itemize{
\item Converts checkbox variables in the dataset from text labels ("Checked" and "Unchecked") to numeric values (0 and 1), and then applies the specified labels.
\item Optionally renames the checkbox variables based on their labels (e.g., transforming variable names like \code{varname___1} to \code{varname_Yes}).
\item Optionally modifies the branching logic in the REDCap dictionary to reflect renamed checkbox options.
}
}
\note{
\itemize{
\item If \code{event_form} is not provided for a longitudinal project, the function may not be able to evaluate branching logic correctly.
}
}
\examples{
# Example with a project object containing data and dictionary
results <- rd_checkbox(project = covican)

# Example with custom labels for the checkboxes
results <- rd_checkbox(
 data = covican$data,
 dic = covican$dictionary,
 checkbox_labels = c("No", "Yes")
)

# Example without renaming checkbox fields
results <- rd_checkbox(covican, checkbox_names = FALSE)

}
