% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_dates.R
\name{rd_dates}
\alias{rd_dates}
\title{Transform Dates and Datetimes in REDCap Data}
\usage{
rd_dates(project = NULL, data = NULL, dic = NULL, event_form = NULL)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{data}{A \code{data.frame} or \code{tibble} representing the REDCap dataset containing the checkbox variables.}

\item{dic}{A \code{data.frame} representing the REDCap dictionary with metadata, including field names, field types, and branching logic.}

\item{event_form}{A \code{data.frame} or \code{list} mapping event names to forms for longitudinal projects. Optional; defaults to \code{NULL} if not applicable.}
}
\value{
A list containing the following elements:
\item{data}{The transformed dataset with date and datetime variables correctly formatted.}
\item{dictionary}{The original data dictionary passed to the function.}
\item{event_form}{The original event-form mapping passed to the function (if applicable).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function processes and transforms date and datetime fields in a REDCap dataset.
It ensures proper handling of data, dictionary (metadata), and event-form mapping,
and applies labels to the dataset for better usability.
}
\details{
The function performs the following tasks:
\itemize{
\item Extracts date and datetime fields from the data dictionary using validation types
(\verb{date_*} and \verb{datetime_*}).
\item Converts these fields in the dataset to \code{Date} and \code{POSIXct} objects, respectively.
}
}
\examples{

# Example usage:
result <- rd_dates(data = covican$data, dic = covican$dictionary)

result <- covican |> rd_dates()

}
