% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_event.R
\name{rd_event}
\alias{rd_event}
\title{Identification of missing event/s}
\usage{
rd_event(
  event,
  filter = NA,
  query_name = NA,
  dic,
  data,
  addTo = NA,
  silent = FALSE,
  report_title = NA,
  report_zeros = FALSE
)
}
\arguments{
\item{event}{Vector with the REDCap's events names to be analyzed.}

\item{filter}{A filter to apply to the dataset. This argument can be used to identify the missing events on a subgroup of the dataset.}

\item{query_name}{Description of the query. It can be defined as the same one for all events or you can define one for each event By default, the function will define the description as `The event [event] is missing' for each event`.}

\item{dic}{R object corresponding to the dictionary of the dataset.}

\item{data}{R object corresponding to the dataset.}

\item{addTo}{Data frame corresponding to a prior report of queries to which you can add the new data frame of queries. By default, the function will always generate a new data frame without taking into account former reports.}

\item{silent}{Logical. If `TRUE`, it does not return a report with the number of queries by event.}

\item{report_title}{Character string with the report's title.}

\item{report_zeros}{Logical. If `TRUE`, it returns a report including events with zero queries.}
}
\value{
A dataframe with 9 columns meant to help the user identify each missing event.
}
\description{
By default, if a record identifier has no information of a designated event, REDCap will not download it.
This function allows you to point out which record identifiers do not have information of a determined event.
}
\keyword{events,}
\keyword{missings}
