% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_export.R
\name{rd_export}
\alias{rd_export}
\title{Exporting Query Dataset}
\usage{
rd_export(
  ...,
  queries = NULL,
  column = NULL,
  sheet_name = NULL,
  path = NULL,
  password = NULL
)
}
\arguments{
\item{...}{List containing the data frame of queries. This list must be the output of the `rd_query` or `rd_event` functions.}

\item{queries}{Data frame containing the identified queries. If the list is specified, this argument is not required.}

\item{column}{Character element specifying the column containing the link for each query.}

\item{sheet_name}{Character element specifying the sheet name of the resulting xlsx file.}

\item{path}{Character element specifying the file path to save the xlsx file. If `NULL`, the file will be created in the current working directory.}

\item{password}{String with the password to protect the worksheet and prevent others from making changes.}
}
\value{
An .xlsx file containing all the queries and, if available, hyperlinks to each of them.
}
\description{
This function exports a query report, generated using the `rd_query` or `rd_event` functions, to an .xlsx file.
}
