% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_insert_na.R
\name{rd_insert_na}
\alias{rd_insert_na}
\title{Insert Missing Values Using a Filter}
\usage{
rd_insert_na(
  project = NULL,
  data = NULL,
  dic = NULL,
  event_form = NULL,
  vars,
  filter
)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{data}{A \code{data.frame} or \code{tibble} representing the REDCap dataset containing the checkbox variables.}

\item{dic}{A \code{data.frame} representing the REDCap dictionary with metadata, including field names, field types, and branching logic.}

\item{event_form}{A \code{data.frame} or \code{list} mapping event names to forms for longitudinal projects. Optional; defaults to \code{NULL} if not applicable.}

\item{vars}{A character vector with the names of the variables to be transformed.}

\item{filter}{A character vector of logical expressions to evaluate. If the evaluation is \code{TRUE}, the corresponding variable in \code{vars} is set to \code{NA}.}
}
\value{
The modified data frame with the specified variables updated.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function allows you to manually insert a missing value into certain variables (\code{vars}) if the specified filter/s (\code{filter}) are satisfied.
It's particularly useful for managing checkboxes without explicit gatekeeper questions in their branching logic.
}
\note{
Each variable is only transformed in the events where both the variable and the filter evaluation are present, so they must have at least one event in common.
}
\examples{

# Example usage:
table(is.na(covican$data$potassium))

data <- covican |>
  rd_insert_na(
    vars = "potassium",
    filter = "age < 65"
  )

table(data$potassium)

}
