% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_query.R
\name{rd_query}
\alias{rd_query}
\title{Identification of Queries}
\usage{
rd_query(
  project = NULL,
  variables = NA,
  expression = NA,
  negate = FALSE,
  event = NA,
  filter = NA,
  addTo = NA,
  variables_names = NA,
  query_name = NA,
  instrument = NA,
  report_title = NA,
  report_zeros = FALSE,
  by_dag = FALSE,
  link = list(),
  data = NULL,
  dic = NULL,
  event_form = NULL
)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{variables}{A character vector containing the names of the variables to be checked for queries.}

\item{expression}{A character vector of expressions to apply to the selected variables.}

\item{negate}{A logical value indicating whether the defined expression should be negated. The default is \code{FALSE}, meaning the expression will be applied as is. If \code{TRUE}, the function will identify values that do \strong{not} meet the condition defined in \code{expression}.}

\item{event}{The name of the REDCap event to analyze. If your REDCap project includes multiple events, you should specify the event to which the variables belong. This is required if your dataset contains multiple events.}

\item{filter}{A character string specifying a filter to be applied to the dataset. This is useful for applying additional conditions, such as using branching logic or filtering based on a specific characteristic of the data (e.g., \code{filter = "available_analytics=='1'"}).}

\item{addTo}{A data frame corresponding to a previous query data frame, to which the new query data frame will be appended. If not provided, the function will create a new data frame for each call. Use this argument to combine multiple queries into one report.}

\item{variables_names}{A character vector containing the descriptions of each selected variable. By default, the function will pull these descriptions from the REDCap dictionary associated with the variables. You can specify custom descriptions if desired.}

\item{query_name}{A character string describing the query. By default, it uses the format \verb{The value is [value] and it should not be [expression]}. You can specify a custom query description for each variable if needed.}

\item{instrument}{The REDCap instrument(s) that the variables belong to. This can be the same for all variables or you can define different instruments for each variable. By default, the function will retrieve the corresponding instrument from the REDCap dictionary.}

\item{report_title}{A character string specifying the title of the final report generated by the function.}

\item{report_zeros}{A logical value indicating whether queries with zero counts should be included in the report. Default is \code{FALSE}. Set it to \code{TRUE} to include variables with zero queries.}

\item{by_dag}{A logical value indicating whether the results should be grouped by Data Access Groups (DAGs). Default is \code{FALSE}. Set to \code{TRUE} to split the results by DAG if applicable.}

\item{link}{A list containing project information to create a web link for each query. This can be used to include clickable links to the REDCap project or other resources directly in the report. If not specified, no links will be included.}

\item{data}{A \code{data.frame} or \code{tibble} representing the REDCap dataset containing the checkbox variables.}

\item{dic}{A \code{data.frame} representing the REDCap dictionary with metadata, including field names, field types, and branching logic.}

\item{event_form}{A \code{data.frame} or \code{list} mapping event names to forms for longitudinal projects. Optional; defaults to \code{NULL} if not applicable.}
}
\value{
A list containing:
\itemize{
\item A data frame of 9 columns (10 columns if \code{link} is specified), providing detailed information on each identified query.
\item A table showing the total number of queries per variable.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function allows you to identify queries based on a specific expression or filter. It is useful for detecting missing values or values that fall outside predefined lower or upper limits of a variable.
The function can also apply branching logic to variables, enabling targeted query identification in REDCap datasets.
}
\examples{
# Example 1: Identifying missing values for multiple variables
example <- rd_query(covican,
  variables = c("copd", "age"),
  expression = c("is.na(x)", "x \%in\% NA"),
  event = "baseline_visit_arm_1"
)
example

# Example 2: Identifying values greater than 20 for the 'age' variable
example <- rd_query(covican,
  variables = "age",
  expression = "x>20",
  event = "baseline_visit_arm_1"
)
example

# Example 3: Identifying missing values for 'potassium' with a filter
example <- rd_query(covican,
  variables = "potassium",
  expression = "is.na(x)",
  event = "baseline_visit_arm_1",
  filter = "available_analytics=='1'"
)
example

}
