% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_recalculate.R
\name{rd_recalculate}
\alias{rd_recalculate}
\title{Recalculate and Verify Calculated Fields in REDCap Data}
\usage{
rd_recalculate(
  project = NULL,
  data = NULL,
  dic = NULL,
  event_form = NULL,
  exclude = NULL
)
}
\arguments{
\item{project}{A list containing the REDCap data, dictionary, and event mapping, typically the output of the \code{redcap_data} function. If provided, it overrides individual \code{data}, \code{dic}, and \code{event_form} arguments.}

\item{data}{A \code{data.frame} or \code{tibble} representing the REDCap dataset containing the checkbox variables.}

\item{dic}{A \code{data.frame} representing the REDCap dictionary with metadata, including field names, field types, and branching logic.}

\item{event_form}{A \code{data.frame} or \code{list} mapping event names to forms for longitudinal projects. Optional; defaults to \code{NULL} if not applicable.}

\item{exclude}{(Optional) A character vector of field names to exclude from recalculation.}
}
\value{
A list containing the following elements:
\item{data}{The updated dataset with recalculated fields (if applicable).}
\item{dictionary}{The updated dictionary with recalculated field entries (if applicable).}
\item{event_form}{The original event-form mapping passed to the function (if applicable).}
\item{results}{A string summarizing the results of the recalculation process.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function processes REDCap project data, recalculates fields defined as calculated fields in the dictionary,
and compares the recalculated values with the original ones. It also generates a report of discrepancies and
updates the dataset and dictionary with new calculated fields (if applicable).
}
\details{
The function:
\itemize{
\item Identifies calculated fields from the dictionary and evaluates the specified formulas.
\item Compares recalculated values with the original values.
\item Adds recalculated fields to the dataset, appending \verb{_recalc} to the original variable names.
\item Updates the dictionary to reflect the new variables.
\item Summarizes the number of calculated fields, discrepancies, and untranslated fields in a report.
}
}
\note{
\itemize{
\item Recalculation is only possible for single-event projects unless \code{event_form} is specified for longitudinal projects.
\item If branching logic is incomplete, poorly defined or contains smart-variables, recalculation may fail for some fields.
}
}
\examples{

# Example usage with individual arguments
results <- rd_recalculate(
  data = covican$data,
  dic = covican$dictionary,
  event_form = covican$event_form
)

# Example usage with a project object, excluding variables from the recalculation
results <- covican |>
  rd_recalculate(exclude = c("age", "screening_fail_crit"))

}
