% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_rlogic.R
\name{rd_rlogic}
\alias{rd_rlogic}
\title{REDCap logic into R logic}
\usage{
rd_rlogic(logic, data)
}
\arguments{
\item{logic}{String containing a logic in REDCap format.}

\item{data}{R object corresponding to the dataset.}
}
\value{
String containing the logic in R format.
}
\description{
This function allows you to transcribe REDCap logic to R logic. WARNING: If the REDCap logic has smart-variables this function will not be able to transform it.
}
\keyword{branching}
\keyword{logic}
\keyword{redcap,}
