% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_transform.R
\name{rd_transform}
\alias{rd_transform}
\title{Transformation of the Raw Data}
\usage{
rd_transform(
  ...,
  data = NULL,
  dic = NULL,
  event_form = NULL,
  checkbox_labels = c("No", "Yes"),
  checkbox_na = FALSE,
  exclude_recalc = NULL,
  exclude_to_factor = NULL,
  delete_vars = NULL,
  delete_pattern = c("_complete", "_timestamp"),
  final_format = "raw",
  which_event = NULL,
  which_form = NULL,
  wide = NULL
)
}
\arguments{
\item{...}{Output of the `redcap_data` function, which is a list containing the data frames of the data, dictionary and event_form (if needed) of the REDCap project.}

\item{data}{Data frame containing the data read from REDCap. If the list is specified, this argument is not necessary.}

\item{dic}{Data frame  containing the dictionary read from REDCap. If the list is specified, this argument is not necessary.}

\item{event_form}{Data frame containing the correspondence of each event with each form. If the list is specified, this argument is not necessary.}

\item{checkbox_labels}{Character vector with the names for the two options of every checkbox variable. Default is `c('No', 'Yes')`.}

\item{checkbox_na}{Logical indicating if checkboxes values with branching logic should be set to missing only when the branching logic is missing (`FALSE`), or also when the branching logic isn't satisfied (`TRUE`). The default is `FALSE`.}

\item{exclude_recalc}{Character vector with the names of variables that should not be recalculated. Useful for projects with time-consuming recalculations of certain calculated fields.}

\item{exclude_to_factor}{Character vector with the names of variables that should not be transformed to factors.}

\item{delete_vars}{Character vector specifying the variables to exclude.}

\item{delete_pattern}{Character vector specifying the regex pattern for variables to be excluded. By default, variables ending with `_complete` and `_timestamp` will be removed.}

\item{final_format}{Character string indicating the final format of the data. Options are `raw`, `by_event` or `by_form`. `raw` (default) returns the transformed data in its original structure, `by_event` returns it as a nested data frame by event, and `by_form` returns it as a nested data frame by form.}

\item{which_event}{Character string indicating a specific event to return if the final format is  `by_event`.}

\item{which_form}{Character string indicating a specific form to return if the final format is `by_form`.}

\item{wide}{Logical indicating if the data split by form (if selected) should be in a wide format (`TRUE`) or a long format (`FALSE`).}
}
\value{
A list with the transformed dataset, dictionary, event_form, and the results of each transformation step.
}
\description{
This function transforms the raw REDCap data read by the `redcap_data` function. It returns the transformed data and dictionary, along with a summary of the results of each step.
}
\examples{
# Basic transformation
rd_transform(covican)

# For customization of checkbox labels (example)
rd_transform(covican,
             checkbox_labels = c("Not present", "Present"))

}
