% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_data.R
\name{redcap_data}
\alias{redcap_data}
\title{Read REDCap data}
\usage{
redcap_data(
  data_path = NA,
  dic_path = NA,
  event_path = NA,
  uri = NA,
  token = NA,
  filter_field = NULL,
  survey_fields = FALSE
)
}
\arguments{
\item{data_path}{Path to the exported R file for data import (if using exported files).}

\item{dic_path}{Path to the dictionary file (CSV or XLSX).}

\item{event_path}{Path to the event-form mapping file (CSV or XLSX) for longitudinal projects (downloadable via the \verb{Designate Instruments for My Events} tab within the \verb{Project Setup} section of REDCap).}

\item{uri}{The URI of the REDCap project (for API connection).}

\item{token}{API token for REDCap project access.}

\item{filter_field}{Fields to include in the import (API connection only).}

\item{survey_fields}{Logical indicating whether to include survey-related fields (API connection only).}
}
\value{
A list containing:
\itemize{
\item \code{data}: Imported dataset.
\item \code{dictionary}: Variable dictionary.
\item \code{event_form} (if applicable): Event-form mapping for longitudinal projects.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function reads datasets from a REDCap project into R for analysis. Data can be imported from REDCap exported files or via an API connection.

\strong{Options for data import:}
\itemize{
\item \strong{Exported Data}: REDCap's \emph{Export Data} function generates files suitable for R import.
\item \strong{API Connection}: Use the REDCap API to directly pull data into R.
}

\strong{Steps for using exported data:}
\enumerate{
\item Use the REDCap \emph{Export Data} function and choose \emph{R Statistical Software} format.
\item REDCap generates:
\itemize{
\item A CSV file with observations.
\item An R script to format variables for import.
}
\item Ensure the exported files, dictionary, and event mapping (if any) are in the same directory.
}
}
\note{
To use other package functions effectively, include the \code{dic_path} argument to load the project dictionary.
}
\examples{
\dontrun{
# Import using exported files

dataset <- redcap_data(
  data_path = "C:/Users/username/example.r",
  dic_path = "C:/Users/username/example_dictionary.csv",
  event_path = "C:/Users/username/events.csv"
)

# Import using API

dataset_api <- redcap_data(
  uri = "https://redcap.idibell.cat/api/",
  token = "55E5C3D1E83213ADA2182A4BFDEA"
) # This token is fictitious
}
}
