% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapDM-package.R
\name{REDCapDM-package}
\alias{REDCapDM-package}
\alias{REDCapDM}
\title{Managing REDCap Data: The R package REDCapDM / REDCapDM: A Toolkit for Managing REDCap Data in R}
\description{
The \strong{REDCapDM} package provides tools to import, process, and manage REDCap data within R.
It supports data retrieval through the REDCap API or directly from exported files and includes a robust
set of functions for data transformation, validation, and discrepancy management. Designed for efficient
workflow integration, \strong{REDCapDM} simplifies the handling of REDCap datasets, making it easier to ensure
data quality and consistency.

Key Features:
\itemize{
\item \strong{Flexible Data Import}: Import data directly from REDCap using API connections or process exported REDCap files.
\item \strong{Data Transformation}: Streamline the cleaning and preparation of raw datasets for analysis.
\item \strong{Query Management}: Identify and track data discrepancies, missing events, and manage resolution reports.
}

Core Functions:
\itemize{
\item \code{redcap_data}: Reads data exported from REDCap or retrieved through the REDCap API.
\item \code{rd_transform}: Processes raw REDCap datasets into a structured and analyzable format.
\itemize{
\item \code{transform_dates}: Transform dates and datetimes variables.
\item \code{recalculate}: Recalculates REDCap calculated fields, compares them to originals, and reports discrepancies.
\item \code{to_factor}: Converts variables to factors and updates the dictionary's branching logic.
\item \code{rd_delete_vars}: Deletes specified or pattern-matched variables from the data and dictionary.
\item \code{transform_checkbox}: Transforms the names of REDCap checkbox variables and updates the branching logic in the dictionary.
\item \code{transform_dic}: Evaluates and transforms branching logic in the REDCap dictionary into R logic.
\item \code{rd_split}: Splits a REDCap dataset by form or event.
}
\item \code{rd_rlogic}: Converts REDCap branching and conditional logic into R-compatible expressions.
\item \code{rd_insert_na}: Inserts missing values into specified variables based on filters.
\item \code{rd_query}: Identifies discrepancies (queries) in the dataset for validation.
\item \code{rd_event}: Detects missing events in longitudinal datasets.
\item \code{check_queries}: Compares historical and current query reports to track changes and additions.
\item \code{rd_export}: Exports a summary report of identified queries to an Excel (.xlsx) file.
}
}
\examples{
\dontrun{
# Install REDCapDM from CRAN:
install.packages("REDCapDM")

# Install the latest version of REDCapDM from GitHub:
remotes::install_github("bruigtp/REDCapDM")
}

}
\keyword{internal}
