% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{split_event}
\alias{split_event}
\title{Creation of a data frame with variables of all the forms of a specified event}
\usage{
split_event(data, dic, event_form, which = NULL)
}
\arguments{
\item{data}{Data frame containing data from REDCap.}

\item{dic}{Data frame  containing the dictionary read from REDCap.}

\item{event_form}{Data frame  containing the correspondence of each event with each form.}

\item{which}{Specify an event if only data for the desired event is wanted.}
}
\description{
Function that given the data, the dictionary and the mapping between forms and events it creates a nested dataset containing all the datasets filtered by each event and containing only the variables found in the event. It can be chosen to return only the data from the specified event.
}
