% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{split_form}
\alias{split_form}
\title{Creation of a Data Frame with Variables from a Specified Form}
\usage{
split_form(data, dic, event_form = NULL, which = NULL, wide = FALSE)
}
\arguments{
\item{data}{Data frame containing data from REDCap.}

\item{dic}{Data frame  containing the dictionary read from REDCap.}

\item{event_form}{Data frame  containing the correspondence of each event with each form.}

\item{which}{Character string specifying a form if only data for that form is desired.}

\item{wide}{Logical indicating if the dataset should be returned in a wide format (`TRUE`) or long format (`FALSE`).}
}
\description{
This function generates a nested dataset containing only the variables associated with each form, using the provided data, dictionary, and event-form mapping. You can choose to return data for a specific form.
}
