% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-transform.R
\name{split_form}
\alias{split_form}
\title{Creation of a data frame with variables of a specified form}
\usage{
split_form(data, dic, event_form = NULL, which = NULL, wide = FALSE)
}
\arguments{
\item{data}{Data frame containing data from REDCap.}

\item{dic}{Data frame  containing the dictionary read from REDCap.}

\item{event_form}{Data frame  containing the correspondence of each event with each form.}

\item{which}{Specify a form if only data for the desired form is wanted.}

\item{wide}{If the dataset needs to be in a wide format or not (long).}
}
\description{
Function that given the data, the dictionary and the mapping between forms and events it creates a nested dataset containing all the datasets having only the variables in each form. It can be chosen to return only the data from the specified form
}
